csa-agent-main/pyproject.toml
[project]
name = "langgraph-cli-exp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.121.3",
    "langchain[openai]>=1.0.7",
    "langgraph>=1.0.2",
    "langgraph-cli[inmem]>=0.4.7",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
]
